Replace the error register with the error mapper
At this moment, all DBus errors are registered with the global error register.
It works fine until you import a third-party module that also uses dasbus and
its error registration. If the registered errors are overlapping, the register
can start to produce unexpected results.

To avoid this problem, it is necessary to remove this global variable from
dasbus. You should create your own instance of the error mapper, assign it to
your message bus and use it for your error registration.

The class ErrorMapper provides the implementation of the error mapper. The
mapper operates with a list of rules that specify the mapping between Python
exception types and DBus error names. The rules are represented by instances
of the classes ErrorRule and DefaultErrorRule.

Call the function get_error_decorator to generate a decorator for DBus errors.