Fix packit archive creation
Dasbus is using setup.py that has two complications for packit archive creation.

- the archive is created in dist/ folder
Packit will create symlink of the tarball into the project root folder. That
should work but for some reason it is not discovered by RPM.
Should be solved by: https://github.com/packit-service/packit/issues/866

- it needs python3-wheel dependency which is not installed on the service (it's
not really needed)
Packit do not want to distribute the archive to PyPi and it's not required
elsewhere. Change the build arguments to remove the metadata and so remove the
dependency.

- overload the reading of the current version
Ideally we should use `upstream_tag_template` with 'v{version}' value, however,
it does not work for some reason.
Should be solved by: https://github.com/packit-service/packit/issues/867