Allow to generate multiple output arguments
The returns_multiple_arguments decorator allows to generate multiple output
arguments in the XML specification of the decorated DBus method. Otherwise,
there will be only one output argument in the specification.

Define a DBus method with multiple output arguments:

    @returns_multiple_arguments
    def Method(self) -> Tuple[Int, Bool]:
        return 0, False

The generated XML specification of the example:

    <method name="Method">
        <arg direction="out" name="return_0" type="i"/>
        <arg direction="out" name="return_1" type="b"/>
    </method>

If the XML specification is not generated by the dbus_interface decorator,
the returns_multiple_arguments decorator has no effect.