Document limitations of the DBus specification generator
It is conventional for member names on DBus to consist of capitalized words
with no punctuation. The generator of the XML specification enforces this
convention to prevent unintended changes in the specification. Some of the
members of the class might not be intended as members of the DBus interface.
Define the __dbus_xml__ attribute, or override the DBusSpecificationGenerator
class as a workaround.